Rename file name _utils to anchor_generator
