+++ src/netspresso_trainer/models/heads/detection/experimental/detection/__init__.py
@@ -1,1 +1,1 @@
-from ._utils import *
+from .anchor_generator import *
